Moved checker strategies into a single file
