--- yin_yang/toggle_checker/checker.py
@@ -1,3 +1,0 @@
-class Checker:
-    def should_be_dark(self) -> bool:
-        raise NotImplementedError('Method should_be_dark() is not implemented')
